requirements.txt
langchain
langchain-community
langchain-groq
groq
tavily-python
pypdf
faiss-cpu
python-dotenv
sentence-transformers
fastapi
uvicorn
